+++ src/packages/@ncigdc/modern_components/GeneExpression/GeneExpressionChart.js
@@ -1,18 +1,10 @@
 import React, { Component } from 'react';
-// import $ from 'jquery';
-// import Konva from 'konva';
-
-// import './inchlib/konva.min';
-import './inchlib/inchlib.css';
-// import InCHlib from './inchlib/inchlib_dev';
 import inchlibData from './inchlib/zhenyu-10x5';
 
-interface IProps {
-  [x:string] :any,
-}
-interface IState {
-  [x:string] :any,
-}
+// ============================= //
+// PSA: don't use inchlib_dev.js //
+// it doesn't work!              //
+// ============================= //
 
 const inchlibOptions = {
   column_metadata: true,
@@ -24,34 +16,24 @@ const inchlibOptions = {
   width: 800,
 };
 
-class GeneExpressionChart extends Component<IProps, IState> {
-  // $inchlib: any;
-  // inchlib: any;
-  // [x:string] :any,
-
-  // inchlib = new InCHlib(inchlibOptions);
+class GeneExpressionChart extends Component {
+  inchlib = new InCHlib(inchlibOptions);
 
   componentDidMount() {
-    const inchlib = new InCHlib(inchlibOptions);
-    inchlib.read_data(inchlibData);
-    inchlib.draw();
-    console.log(inchlib);
-
-    // this.inchlib = inchlib;
+    this.inchlib.read_data(inchlibData);
+    this.inchlib.draw();
   }
-  // componentWillUnmount() {
+
+  componentWillUnmount() {
     // if inchlib doesn't have cleanup methods,
     // we have to add our own.
-    // TODO remove event listeners
-    // this.inchlib.destroy();
-  // }
+    // TODO remove event listeners added in componentDidMount
+    this.inchlib.destroy();
+  }
 
   render() {
     return (
-      <div
-        id="inchlib"
-        // ref={inchlib => this.inchlib = inchlib}
-        />
+      <div id="inchlib" />
     );
   }
 }
